Fixed broken PID code. Replaced by a different algorithm.
Also did some raw tuning of the PID according to the new algorithm

+++ PID/LogValue.cs
@@ -6,7 +6,7 @@ namespace BrewMatic3000.PID
     public class LogValue
     {
         public DateTime TimeStamp { get; set; }
-        public float Temperature { get; set; }
-        public float Effect { get; set; }
+        public double Temperature { get; set; }
+        public double Effect { get; set; }
     }
 }
